requirements.txt
asgiref==3.5.1
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.5
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==14.1.1
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
